edit transactions api, add register endpoint

+++ BankApp/Shared/Entities/Category.cs
@@ -20,7 +20,7 @@ namespace BankApp.Shared.Entities
         [FirestoreProperty]
         public string OwnerId { get; set; }
 
-        public decimal TransactionSum { get; set; }
+        public double TransactionSum { get; set; }
         public bool IsBeingModified { get; set; }
     }
 }
